Hack to preserve notification after adding post
Notifications are getting cleared due to the redirect between
"post_page" and "post_page_with_slug".  Exclude these from the routes
that clear notifications, so that we are ensured of always seeing the
after-post-adding notification.

Downside: if there were a notification from somewhere else, it wouldn't
get cleared when we navigate to a post.