Fix usage of ``can.edit`` for users
Fix the ``changeEmail`` method's invocation signature of ``can.edit``,
and fix the definition of ``can.edit`` to allow it to work with users as
the "item".

Note that this relies on "_id" to be unique across collections.  Mongo
won't generate the same ID in two collections, but it is possible to
manually add an object to another collection with a duplicate ID.